Updated the list of configuration settings and fixed some typos in the comments.
Also updated the generation of the NuttX template code. Only the bare minimum is copied into the source directory (Only the configuration files for the NuttX project and application, along with the template source files if requested). The rest of the RTOS code is merged at the build stage in a separate function.


Former-commit-id: 501b0a4eaa1af574360b92518f4f3ba0309bbd3a [formerly 5aab9ede0121dbae368b949c1531510a834a948a] [formerly 0ebe807f582160eadf0a2648a45da9c717a46a72 [formerly e2fbe7133844c37443021bd9b6de1f9d235b9548]]
Former-commit-id: a03ea94a0bc1550e1a1d3b2ef13bc153d56fc216 [formerly d30dc69f0a77a173e1bdff2207707bc03eac6416]
Former-commit-id: cb8f2ef635693fcbe4fa329e26dcd7f6f5149f46