Modified the if statements which determine which template file to create. Now it checks whether four variables are defined (TEMPLATE_C_SOURCE, TEMPLATE_CPP_SOURCE, TEMPLATE_H_SOURCE and TEMPLATE_HPP_SOURCE, which should all defined in the build_config). If these variables are present, it uses them to find the template files. If not, the old if statements are used to determine where to look.
The new check also includes a variable ${TEMPLATE_SUFFIX} to append a suffix onto the template file names (NuttX requires the main file to have the structure <FILENAME>_main.c). ${TEMPLATE_SUFFIX} is also defined in the build_config file, but is not necessary, as it will otherwise expand into an empty string.

Finally, another instance of [ $PLATFORM = "freertos" ] was fixed.


Former-commit-id: 4df8ae5c79a414ccaecfa85c4718ab2bab567c5e [formerly 108a3ca6d404d9627e9c86a5c2cf42938048e848] [formerly 1a4277615bd37c474b7da727f913f446b49ae871 [formerly ee2d3e3f61cde4b7519fbd7ae10bfe9134ba71d0]]
Former-commit-id: f2b9dfb9d87b6e488b2a802b1b337c0e74351f1f [formerly b67b151686614376b1c9a156a92845daa59d4004]
Former-commit-id: aea0d7c5d61b29e83ce646af88f7c624cae13cb0