Updated create_component to check if having mutliple executable files is allowed (By checking if the "MULTIPLE_EXECUTABLES" variable is defined). If so, it now prompts the user to enter a string of executable file names and adds them to the .cfg file. In the event that the user chooses not to, or the project does not qualify for multiple executables, the "EXECUTABLES" key is still created in the component .cfg file but it is assigned the value of "COMPONENT_NAME" instead.
If the user requests template files be created, files for each executable are also created, and the user is asked whether a C or CPP file should be created for each executable. The name of the executable is also printed when deciding which type of template file to generate to minimise user confusion.

The STM32F103RBT6_NuttX build config was also updated to test the creation of multiple exectuable files.


Former-commit-id: 05addc04bd0706670956f30ee1fd5ee8e47eb381 [formerly eca0fca24d0d54213deda8acc1e34e4c17877223] [formerly 94c30a003f705d73588890d7ac75af69e5fd5dac [formerly 206f5231c8de53e5bc805242844abd88a0dfc26f]]
Former-commit-id: 2a17cabc4ddea6eb68eb14c596ae75bdec5f5f26 [formerly 2b4ae2d7249bd385f4b5b0409e9f3533e86f6f14]
Former-commit-id: 823b5987ee265623d8f53cb6f87173731fb58060